pyproject.toml
[tool.pyright]
reportUninitializedInstanceVariable = false
reportUnannotatedClassAttribute = false

[project]
name = "sharing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aerich[toml]>=0.8.2",
  "bcrypt>=4.3.0",
  "cutword>=0.1.1",
  "elasticsearch[async]>=9.0.1",
  "email-validator>=2.2.0",
  "fastapi[all]>=0.115.12",
  "granian>=2.2.6",
  "httpx>=0.28.1",
  "passlib>=1.7.4",
  "python-jose>=3.4.0",
  "tortoise-orm[asyncpg]>=0.25.0",
]
